rootless: move ./rootless to ./pkg/rootless
(cherry picked from commit 6ab12ec8f48080b521020518e8632868fd6cff96)
